refactored Oswap adapter by extracting common reducer function that sums up TVL of multiple fetch calls into Obyte helpers
This is a preparation for adding more Obyte adapters that will calculate
TVL similarly to Oswap.

No functional change in this commit.
